+++ com/bridgelabz/LinkedList.java
@@ -25,9 +25,22 @@ public class LinkedList {
 			temp = temp.next;
 		}
 	}
-}
+	
+	public void append(int data) {
+		Node newNode = new Node(data);
+		if(head == null) {
+			head = newNode;
+			tail = newNode;
+			return;
+		}
+		else {
+			tail.next = newNode;
+			tail = newNode;
+		}	
+	}
 
-class Node {
+}
+ class Node {
 	
 	int data;
 	Node next;
@@ -37,3 +50,4 @@ class Node {
 		this.next = null;
 	}
 }
+
