--- src/com/bridgelabz/LinkedListMain.java
@@ -1,12 +1,0 @@
-package com.bridgelabz;
-public class LinkedListMain {
-	
-public static void main(String[] args) {
-		System.out.println("welcome to linked list program");
-		LinkedList list = new LinkedList();
-		list.add(56);
-		list.add(30);
-		list.add(70);
-		list.show();
-	}
-}
